Address stream telemetry review feedback
Poll telemetry only while its WebRTC panel is open, extract its renderer, remove duplicate Android FPS, and document vendored runtime dependencies and route ownership.
