+++ .changeset/webrtc-stream-statistics.md
@@ -2,4 +2,4 @@
 'expo-device-hub': minor
 ---
 
-Add WebRTC stream statistics.
+Add WebRTC stream statistics. Move Android network and font-scale compatibility routes to serve-emu.
